Fix CI

+++ esphome/core/time.cpp
@@ -71,10 +71,11 @@ bool ESPTime::strptime(const std::string &time_to_parse, ESPTime &esp_time) {
   uint8_t hour;
   uint8_t minute;
   uint8_t second;
-  uint8_t num;
+  int num;
 
-  if (sscanf(time_to_parse.c_str(), "%04" SCNu16 "-%02" SCNu8 "-%02" SCNu8 " %02" SCNu8 ":%02" SCNu8 ":%02" SCNu8 " %n",
-             &year, &month, &day, &hour, &minute, &second, &num) == 6 &&
+  if (sscanf(time_to_parse.c_str(),
+             "%04" SCNu16 "-%02" SCNu8 "-%02" SCNu8 " %02" SCNu8 ":%02" SCNu8 ":%02" SCNu8 " %n",  // NOLINT
+             &year, &month, &day, &hour, &minute, &second, &num) == 6 &&                           // NOLINT
       num == time_to_parse.size()) {
     esp_time.year = year;
     esp_time.month = month;
@@ -82,18 +83,20 @@ bool ESPTime::strptime(const std::string &time_to_parse, ESPTime &esp_time) {
     esp_time.hour = hour;
     esp_time.minute = minute;
     esp_time.second = second;
-  } else if (sscanf(time_to_parse.c_str(), "%02" SCNu8 ":%02" SCNu8 ":%02" SCNu8 " %n", &hour, &minute, &second,
-                    &num) == 3 &&
+  } else if (sscanf(time_to_parse.c_str(), "%02" SCNu8 ":%02" SCNu8 ":%02" SCNu8 " %n", &hour, &minute,  // NOLINT
+                    &second,                                                                             // NOLINT
+                    &num) == 3 &&                                                                        // NOLINT
              num == time_to_parse.size()) {
     esp_time.hour = hour;
     esp_time.minute = minute;
     esp_time.second = second;
-  } else if (sscanf(time_to_parse.c_str(), "%02" SCNu8 ":%02" SCNu8 " %n", &hour, &minute, &num) == 2 &&
+  } else if (sscanf(time_to_parse.c_str(), "%02" SCNu8 ":%02" SCNu8 " %n", &hour, &minute, &num) == 2 &&  // NOLINT
              num == time_to_parse.size()) {
     esp_time.hour = hour;
     esp_time.minute = minute;
     esp_time.second = 0;
-  } else if (sscanf(time_to_parse.c_str(), "%04" SCNu16 "-%02" SCNu8 "-%02" SCNu8 " %n", &year, &month, &day, &num) ==
+  } else if (sscanf(time_to_parse.c_str(), "%04" SCNu16 "-%02" SCNu8 "-%02" SCNu8 " %n", &year, &month, &day,  // NOLINT
+                    &num) ==                                                                                   // NOLINT
                  3 &&
              num == time_to_parse.size()) {
     esp_time.year = year;
